BREAKING_CHANGE: move execute_{sync,async} methods to SolvedDependent
There are backwards compatibility shims on Container.execute_{sync,async} so this should, in theory, not break any code, but it is nontheless a pretty large API change.

This is (hopefully, time permitting) going to be a part of a larger effort to simplify the APIs and implementation of this library
