+++ docs_src/async_init_dependency.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from di.container import Container
+from di import Container
 from di.dependent import Dependent, Marker
 from di.executors import AsyncExecutor
 from di.typing import Annotated
@@ -24,5 +24,5 @@ async def main() -> None:
     executor = AsyncExecutor()
     solved = container.solve(Dependent(endpoint), scopes=(None,))
     async with container.enter_scope(None) as state:
-        res = await container.execute_async(solved, executor=executor, state=state)
+        res = await solved.execute_async(executor=executor, state=state)
         assert res == "👋"
