+++ docs_src/autowiring.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from di.container import Container
+from di import Container
 from di.dependent import Dependent
 from di.executors import AsyncExecutor
 
@@ -23,4 +23,4 @@ async def framework():
     container = Container()
     solved = container.solve(Dependent(endpoint, scope="request"), scopes=["request"])
     async with container.enter_scope("request") as state:
-        await container.execute_async(solved, executor=AsyncExecutor(), state=state)
+        await solved.execute_async(executor=AsyncExecutor(), state=state)
