+++ docs_src/markers.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from di.container import Container
+from di import Container
 from di.dependent import Dependent, Marker
 from di.executors import SyncExecutor
 from di.typing import Annotated
@@ -28,4 +28,4 @@ def framework():
     container = Container()
     solved = container.solve(Dependent(endpoint, scope="request"), scopes=["request"])
     with container.enter_scope("request") as state:
-        container.execute_sync(solved, executor=SyncExecutor(), state=state)
+        solved.execute_sync(executor=SyncExecutor(), state=state)
