+++ docs_src/simple.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 
-from di.container import Container
+from di import Container
 from di.dependent import Dependent
 from di.executors import SyncExecutor
 
@@ -23,7 +23,7 @@ def main():
     container = Container()
     solved = container.solve(Dependent(C, scope="request"), scopes=["request"])
     with container.enter_scope("request") as state:
-        c = container.execute_sync(solved, executor=SyncExecutor(), state=state)
+        c = solved.execute_sync(executor=SyncExecutor(), state=state)
     assert isinstance(c, C)
     assert isinstance(c.a, A)
     assert isinstance(c.b, B)
