+++ docs_src/singleton.py
@@ -1,6 +1,6 @@
 import inspect
 
-from di.container import Container
+from di import Container
 from di.dependent import Dependent
 from di.executors import SyncExecutor
 
@@ -23,7 +23,7 @@ def framework():
     executor = SyncExecutor()
     with container.enter_scope("app") as app_state:
         with container.enter_scope("request", state=app_state) as req_state:
-            repo1 = container.execute_sync(solved, executor=executor, state=req_state)
+            repo1 = solved.execute_sync(executor=executor, state=req_state)
         with container.enter_scope("request", state=app_state) as req_state:
-            repo2 = container.execute_sync(solved, executor=executor, state=req_state)
+            repo2 = solved.execute_sync(executor=executor, state=req_state)
         assert repo1 is repo2
