+++ tests/test_task.py
@@ -6,7 +6,7 @@ from typing import Any, AsyncGenerator, Callable, Generator
 
 import pytest
 
-from di.container import Container
+from di import Container
 from di.dependent import Dependent
 from di.executors import ConcurrentAsyncExecutor
 
@@ -92,13 +92,13 @@ async def test_dependency_types(
     solved = container.solve(Dependent(dep, use_cache=use_cache), scopes=[None])  # type: ignore
     executor = ConcurrentAsyncExecutor()
     async with container.enter_scope(None) as state:
-        res = await container.execute_async(solved, executor=executor, state=state)
+        res = await solved.execute_async(executor=executor, state=state)
         assert res == 1
         # test the cached execution paths
-        res = await container.execute_async(solved, executor=executor, state=state)
+        res = await solved.execute_async(executor=executor, state=state)
         assert res == 1
         # test the by value execution paths
-        res = await container.execute_async(
-            solved, executor=executor, values={dep: 2}, state=state
+        res = await solved.execute_async(
+            executor=executor, values={dep: 2}, state=state
         )
         assert res == 2
