+++ tests/test_wiring_annotated.py
@@ -1,6 +1,6 @@
 from typing import Optional, Tuple
 
-from di.container import Container, bind_by_type
+from di import Container, bind_by_type
 from di.dependent import Dependent, Marker
 from di.executors import SyncExecutor
 from di.typing import Annotated
@@ -45,7 +45,7 @@ def test_autowiring_class_with_default_builtin() -> None:
     solved = container.solve(dep, scopes=[None])
 
     with container.enter_scope(None) as state:
-        injected_values = container.execute_sync(solved, SyncExecutor(), state=state)
+        injected_values = solved.execute_sync(SyncExecutor(), state=state)
 
     assert injected_values == ("default", 1)
 
@@ -67,7 +67,7 @@ def test_autowiring_class_with_default_class() -> None:
     solved = container.solve(dep, scopes=[None])
 
     with container.enter_scope(None) as state:
-        injected_value = container.execute_sync(solved, SyncExecutor(), state=state)
+        injected_value = solved.execute_sync(SyncExecutor(), state=state)
 
     assert injected_value == "default"
 
@@ -90,6 +90,6 @@ def test_autowiring_class_with_default_class_from_bind() -> None:
     solved = container.solve(dep, scopes=[None])
 
     with container.enter_scope(None) as state:
-        injected_value = container.execute_sync(solved, SyncExecutor(), state=state)
+        injected_value = solved.execute_sync(SyncExecutor(), state=state)
 
     assert injected_value == "bound"
